move: domain 하위 repository 디렉토리로 이동
